server/requirements.txt
# Runtime dependencies for the server container
# These are pinned for reproducibility inside Docker.
# The pyproject.toml is the source of truth for local dev.

openenv-core[core]>=0.2.2
pytest>=8.0.0
pytest-json-report>=1.5.0
flake8>=7.0.0
openai>=1.0.0
uvicorn[standard]>=0.29.0
